requirements.txt
beautifulsoup4
solara
python-dotenv
solara
pandas
elasticsearch
protobuf==3.20.0
langchain
langchain-community
tqdm
openai
transformers
torch
sentence-transformers
bitsandbytes
accelerate
evaluate
rouge_score
bert_score
setuptools
sentencepiece
streamlit